refactor(plugin): allow cycle reference for `PluginDriver`
