+++ crates/rolldown/Cargo.toml
@@ -31,7 +31,6 @@ rolldown_sourcemap = { workspace = true }
 rolldown_tracing   = { workspace = true }
 rolldown_utils     = { workspace = true }
 rustc-hash         = { workspace = true }
-smallvec           = { workspace = true }
 string_wizard      = { workspace = true }
 sugar_path         = { workspace = true }
 tokio              = { workspace = true, features = ["rt", "macros", "sync"] }
