+++ crates/rolldown/src/utils/render_chunks.rs
@@ -1,9 +1,9 @@
-use rolldown_common::RenderedChunk;
+use rolldown_common::{IntoBatchedResult, RenderedChunk};
 use rolldown_plugin::{HookRenderChunkArgs, SharedPluginDriver};
 use rolldown_sourcemap::SourceMap;
 use rolldown_utils::block_on_spawn_all;
 
-use crate::error::{BatchedErrors, IntoBatchedResult};
+use crate::error::BatchedErrors;
 
 pub async fn render_chunks<'a>(
   plugin_driver: &SharedPluginDriver,
