+++ crates/rolldown_common/src/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod ast_scope;
+pub mod batched_errors;
 pub mod chunk_id;
 pub mod chunk_kind;
 pub mod entry_point;
