+++ crates/rolldown_plugin/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 
 [dependencies]
 async-trait        = { workspace = true }
+futures            = { workspace = true }
 rolldown_common    = { workspace = true }
 rolldown_error     = { workspace = true }
 rolldown_sourcemap = { workspace = true }
