+++ crates/rolldown_plugin/src/plugin_context.rs
@@ -1,0 +1,8 @@
+use std::sync::Weak;
+
+use crate::PluginDriver;
+
+#[derive(Debug, Default)]
+pub struct PluginContext {
+  pub(crate) _plugin_driver: Weak<PluginDriver>,
+}
